Renamespace footer button
